+++ handlers/accounts_handler.go
@@ -11,6 +11,7 @@ import (
 	"github.com/go-co-op/gocron"
 	log "github.com/sirupsen/logrus"
 	"github.com/tailwarden/komiser/models"
+	"github.com/tailwarden/komiser/repository"
 	"github.com/tailwarden/komiser/utils"
 	"github.com/uptrace/bun"
 	"github.com/uptrace/bun/dialect/pgdialect"
@@ -38,7 +39,7 @@ func (handler *ApiHandler) IsOnboardedHandler(c *gin.Context) {
 
 	accounts := make([]models.Account, 0)
 
-	_, err := models.HandleQuery(handler.ctx, handler.db, "LIST", &accounts, nil)
+	_, err := repository.HandleQuery(handler.ctx, handler.db, "LIST", &accounts, nil)
 	if err != nil {
 		log.WithError(err).Error("scan failed")
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "scan failed"})
@@ -62,7 +63,7 @@ func (handler *ApiHandler) ListCloudAccountsHandler(c *gin.Context) {
 		return
 	}
 
-	_, err := models.HandleQuery(handler.ctx, handler.db, "LIST", &accounts, nil)
+	_, err := repository.HandleQuery(handler.ctx, handler.db, "LIST", &accounts, nil)
 	if err != nil {
 		log.WithError(err).Error("scan failed")
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "scan failed"})
@@ -74,7 +75,7 @@ func (handler *ApiHandler) ListCloudAccountsHandler(c *gin.Context) {
 			Total int `bun:"total" json:"total"`
 		}{}
 
-		_, err := models.HandleQuery(handler.ctx, handler.db, "RESOURCE_COUNT", &output, [][3]string{{"provider", "=", account.Provider}, {"account", "=", account.Name}})
+		_, err := repository.HandleQuery(handler.ctx, handler.db, "RESOURCE_COUNT", &output, [][3]string{{"provider", "=", account.Provider}, {"account", "=", account.Name}})
 		if err != nil {
 			fmt.Println(err)
 			c.JSON(http.StatusInternalServerError, gin.H{"error": "scan failed"})
@@ -107,7 +108,7 @@ func (handler *ApiHandler) NewCloudAccountHandler(c *gin.Context) {
 		unsavedAccounts = append(unsavedAccounts, account)
 	} else {
 
-		result, err := models.HandleQuery(handler.ctx, handler.db, "INSERT", &account, nil)
+		result, err := repository.HandleQuery(handler.ctx, handler.db, "INSERT", &account, nil)
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 			return
@@ -155,7 +156,7 @@ func (handler *ApiHandler) ReScanAccount(c *gin.Context) {
 
 	account := new(models.Account)
 	account.Status = "SCANNING"
-	res, err := models.HandleQuery(handler.ctx, handler.db, "RE_SCAN_ACCOUNT", account, [][3]string{{"id", "=", accountId}, {"status", "=", "CONNECTED"}})
+	res, err := repository.HandleQuery(handler.ctx, handler.db, "RE_SCAN_ACCOUNT", account, [][3]string{{"id", "=", accountId}, {"status", "=", "CONNECTED"}})
 	if err != nil {
 		log.Error("Couldn't set status", err)
 		return
@@ -172,7 +173,7 @@ func (handler *ApiHandler) DeleteCloudAccountHandler(c *gin.Context) {
 	accountId := c.Param("id")
 
 	account := new(models.Account)
-	_, err := models.HandleQuery(handler.ctx, handler.db, "DELETE", account, [][3]string{{"id", "=", accountId}})
+	_, err := repository.HandleQuery(handler.ctx, handler.db, "DELETE", account, [][3]string{{"id", "=", accountId}})
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -191,7 +192,7 @@ func (handler *ApiHandler) UpdateCloudAccountHandler(c *gin.Context) {
 		return
 	}
 
-	_, err = models.HandleQuery(handler.ctx, handler.db, "UPDATE_ACCOUNT", &account, [][3]string{{"id", "=", accountId}})
+	_, err = repository.HandleQuery(handler.ctx, handler.db, "UPDATE_ACCOUNT", &account, [][3]string{{"id", "=", accountId}})
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
