+++ handlers/alerts_handler.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/tailwarden/komiser/models"
+	"github.com/tailwarden/komiser/repository"
 )
 
 func (handler *ApiHandler) IsSlackEnabledHandler(c *gin.Context) {
@@ -33,7 +34,7 @@ func (handler *ApiHandler) NewAlertHandler(c *gin.Context) {
 		return
 	}
 
-	result, err := models.HandleQuery(handler.ctx, handler.db, "INSERT", &alert, nil)
+	result, err := repository.HandleQuery(handler.ctx, handler.db, "INSERT", &alert, nil)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -62,7 +63,7 @@ func (handler *ApiHandler) UpdateAlertHandler(c *gin.Context) {
 		return
 	}
 
-	_, err = models.HandleQuery(handler.ctx, handler.db, "UPDATE_ALERT", &alert, [][3]string{{"id", "=", fmt.Sprint(alertId)}})
+	_, err = repository.HandleQuery(handler.ctx, handler.db, "UPDATE_ALERT", &alert, [][3]string{{"id", "=", fmt.Sprint(alertId)}})
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -75,7 +76,7 @@ func (handler *ApiHandler) DeleteAlertHandler(c *gin.Context) {
 	alertId := c.Param("id")
 
 	alert := new(models.Alert)
-	_, err := models.HandleQuery(handler.ctx, handler.db, "DELETE", alert, [][3]string{{"id", "=", fmt.Sprint(alertId)}})
+	_, err := repository.HandleQuery(handler.ctx, handler.db, "DELETE", alert, [][3]string{{"id", "=", fmt.Sprint(alertId)}})
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
